+++ src/Recyclarr.TrashLib/Repo/VersionControl/IGitRepository.cs
@@ -6,6 +6,6 @@ public interface IGitRepository : IDisposable
     Task Fetch(string remote = "origin");
     Task ResetHard(string toBranchOrSha1);
     Task SetRemote(string name, Uri newUrl);
-    Task Clone(Uri cloneUrl, string? branch = null);
+    Task Clone(Uri cloneUrl, string? branch = null, int depth = 0);
     Task Status();
 }