+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Reduce the time it takes to clone the config and trash repositories by performing shallow clones
+  (#201).
+
 ## [5.2.0] - 2023-08-06
 
 ### Added